Added additional functionalities to gamemap.
You can now request the tile position when giving the tile sprite.
Getting the tile position natively will give you the left bottom corner,
whereas the other sprites the position will give you the center of it.

The tileToWorld will scale the position (map scale) and return the
position with respect to the center point.