Added event handlers.
These event handlers are currently actived when a player tile position
is on a tile with an event tile. An event tile is a tile in the layer
event.

Each event tile has a property "event" which a integer representing one
of the tile properties as defined in GameMap.h.