Wrote the save/load config part
It now saves/loads the CONFIG file. The config file currently only
supports the chapter and spawnpoint. It will successfull set the
spawnpoint and chapter in the progress attribute of ChapterManager.

In the future it should also support which achievements/unlocks/dialogs
the player has. This should make debugging easier.

I had to choose between xml/json and chose for xml because the other
configuration files are either plist (which is xml) and xml.
Another choose was either to use the ValueMap of cocos2d or write my own
structure. I choose for my own structure because a ValueMap is hard to
debug and makes it more complicated. The downside with this is that I
have to write my own xml<->struct.