Smoothed out the collision vertical and horizontal
It would sometimes perform a double jump when you do a strafe jump to
ledge. This was due the fact that It would resolve it as a vertical
collision, push the character up and jump again. Also horizontal
collision followed by a vertical collision made the fall sometimes
awkward.

Fixed by setting the min of 0 or velocity.y with horizontal collision
(fixed the double jump)

And by setting the min/max of 0 or velocity.y in a vertical collision by
top or bottom respectively.