Started on working on parsing the dialog xml files
It is now a hacky code. But the idea is to work it out in the
chapter1.cpp and then move all the parsing to DíalogAction. I putted
everthing here just that is easier for now.

It now successfull parses move and text. And holy shit this makes making
new dialogs easier.