Started to rework on the architecture: physics
Splitted the physics and collision detection from the basicscene.cpp

The architecture will now be component based. The player whill have for
example a physic component and an animation component.
The reason is that it separates concern and makes it easier to extend
the game. Currently that is the biggest problem, too many concers are
placed into a single entity.